+++ core/capabilities/integration_tests/framework/don.go
@@ -37,6 +37,22 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/utils/testutils/heavyweight"
 )
 
+type DonContext struct {
+	EthBlockchain      *EthBlockchain
+	p2pNetwork         *MockRageP2PNetwork
+	capabilityRegistry *CapabilitiesRegistry
+}
+
+func CreateDonContext(ctx context.Context, t *testing.T) DonContext {
+	ethBlockchain := NewEthBlockchain(t, 1000, 1*time.Second)
+	rageP2PNetwork := NewMockRageP2PNetwork(t, 1000)
+	capabilitiesRegistry := NewCapabilitiesRegistry(ctx, t, ethBlockchain)
+
+	servicetest.Run(t, rageP2PNetwork)
+	servicetest.Run(t, ethBlockchain)
+	return DonContext{EthBlockchain: ethBlockchain, p2pNetwork: rageP2PNetwork, capabilityRegistry: capabilitiesRegistry}
+}
+
 type capabilityNode struct {
 	*cltest.TestApplication
 	registry  *capabilities.Registry
@@ -362,4 +378,8 @@ func addOCR3Capability(ctx context.Context, t *testing.T, lggr logger.Logger, ca
 	transmitter := ocr3.NewContractTransmitter(lggr, capabilityRegistry, "")
 
 	libocr.AddNode(plugin, transmitter, ocr2KeyBundle)
+}
+
+func Context(tb testing.TB) context.Context {
+	return testutils.Context(tb)
 }
